feat(cb2-11176): refactor plate endpoint to simplify validation
