+++ tests/unit/validators/plate.unit.test.ts
@@ -3,7 +3,7 @@ const mockValidateSysNumTimestampPathParams = jest.fn();
 
 import { APIGatewayProxyEvent } from 'aws-lambda';
 import { PlateReasonForIssue, Plates } from '../../../src/models/plate';
-import { validatePlateErrors, validatePlateInfo } from '../../../src/validators/plate';
+import { validatePlateInfo, validatePlateRequestBody } from '../../../src/validators/plate';
 
 jest.mock('../../../src/validators/sysNumTimestamp', () => ({
   validateSysNumTimestampPathParams: mockValidateSysNumTimestampPathParams,
@@ -16,37 +16,39 @@ describe('Test plate validator', () => {
         statusCode: 400,
         body: 'Missing system number',
       });
-      const res = validatePlateErrors({} as unknown as APIGatewayProxyEvent);
+      const res = validatePlateRequestBody({} as unknown as APIGatewayProxyEvent);
       expect(res).toEqual({ statusCode: 400, body: 'Missing system number' });
     });
 
     it('should error if there is an empty body', () => {
       mockValidateSysNumTimestampPathParams.mockReturnValueOnce(undefined);
-      const res = validatePlateErrors({} as unknown as APIGatewayProxyEvent);
+      const res = validatePlateRequestBody({} as unknown as APIGatewayProxyEvent);
       expect(res).toEqual({ statusCode: 400, body: 'invalid request' });
     });
 
     it('should error if there is no reason for creation', () => {
       mockValidateSysNumTimestampPathParams.mockReturnValueOnce(undefined);
-      const res = validatePlateErrors({ body: JSON.stringify({ foo: 'bar' }) } as unknown as APIGatewayProxyEvent);
+      const res = validatePlateRequestBody({ body: JSON.stringify({ foo: 'bar' }) } as unknown as APIGatewayProxyEvent);
       expect(res).toEqual({ statusCode: 400, body: 'Reason for creation not provided' });
     });
 
     it('should error if there is no vtmUsername', () => {
       mockValidateSysNumTimestampPathParams.mockReturnValueOnce(undefined);
-      const res = validatePlateErrors({ body: JSON.stringify({ reasonForCreation: 'bar' }) } as unknown as APIGatewayProxyEvent);
+      const res = validatePlateRequestBody({ body: JSON.stringify({ reasonForCreation: 'bar' }) } as unknown as APIGatewayProxyEvent);
       expect(res).toEqual({ statusCode: 400, body: 'No username provided' });
     });
 
     it('should error if there is no recipientEmailAddress', () => {
       mockValidateSysNumTimestampPathParams.mockReturnValueOnce(undefined);
-      const res = validatePlateErrors({ body: JSON.stringify({ reasonForCreation: 'bar', vtmUsername: 'foo' }) } as unknown as APIGatewayProxyEvent);
+      const res = validatePlateRequestBody(
+        { body: JSON.stringify({ reasonForCreation: 'bar', vtmUsername: 'foo' }) } as unknown as APIGatewayProxyEvent,
+      );
       expect(res).toEqual({ statusCode: 400, body: 'No recipient email provided' });
     });
 
     it('should pass and return undefined', () => {
       mockValidateSysNumTimestampPathParams.mockReturnValueOnce(undefined);
-      const res = validatePlateErrors({
+      const res = validatePlateRequestBody({
         body: JSON.stringify({
           reasonForCreation: 'bar',
           vtmUsername: 'foo',
